+++ CHANGELOG.md
@@ -185,6 +185,7 @@
 - [Short-hand syntax for `order_by` added.][3643]
 - [Expanded `Table.at` to support index access and added `Table.column_count`
   method.][3644]
+- [Removed `Array.set_at`.][3634]
 
 [debug-shortcuts]:
   https://github.com/enso-org/enso/blob/develop/app/gui/docs/product/shortcuts.md#debug
@@ -292,6 +293,7 @@
 [3648]: https://github.com/enso-org/enso/pull/3648
 [3665]: https://github.com/enso-org/enso/pull/3665
 [5250]: https://github.com/enso-org/enso/pull/5250
+[3634]: https://github.com/enso-org/enso/pull/3634
 
 #### Enso Compiler
 
@@ -372,7 +374,7 @@
 [3631]: https://github.com/enso-org/enso/pull/3631
 [3633]: https://github.com/enso-org/enso/pull/3633
 [3637]: https://github.com/enso-org/enso/pull/3637
-[3633]: https://github.com/enso-org/enso/pull/3641
+[3641]: https://github.com/enso-org/enso/pull/3641
 [3658]: https://github.com/enso-org/enso/pull/3658
 
 # Enso 2.0.0-alpha.18 (2021-10-12)